% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{main}
\alias{main}
\title{Run the data cleaning pipeline from command line}
\usage{
main(input_file_path, output_file_path)
}
\arguments{
\item{input_file_path}{Path to input CSV file}

\item{output_file_path}{Path to save output CSV file}
}
\value{
Writes a cleaned CSV file to disk
}
\description{
Run the data cleaning pipeline from command line
}
\examples{
\dontrun{
main("raw_data.csv", "cleaned_data.csv")
}
}
