% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Create a table using the given data columns}
\usage{
make_table(data_cols, folder, name)
}
\arguments{
\item{data_cols}{A dataframe or individual column(s) of a dataframe}

\item{folder}{The folder and path to write the file to}

\item{name}{The name of the file to be written}
}
\value{
Writes a new CSV file
}
\description{
Create a table using the given data columns
}
\examples{
\dontrun{
make_table(mtcars$mpg, "output", "mpg_table.csv")
}
}
