% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_barplot.R
\name{make_barplot}
\alias{make_barplot}
\title{Generate a Proportional Stacked Bar Plot}
\usage{
make_barplot(dataset, x, x_name, class, class_name)
}
\arguments{
\item{dataset}{A data frame containing the data to be plotted.}

\item{x}{A string specifying the name of the categorical variable to be plotted on the x-axis.}

\item{x_name}{A string used as the label for the x-axis.}

\item{class}{A string specifying the name of the grouping variable to fill the bars.}

\item{class_name}{A string used as the label for the y-axis (often something like "Proportion" or the name of the group variable).}
}
\value{
A ggplot object showing the proportional stacked bar chart.
}
\description{
Generate a Proportional Stacked Bar Plot
}
\examples{
\dontrun{
data <- data.frame(
  animal_type = c("Dog", "Dog", "Cat", "Cat", "Dog"),
  outcome_group = c("Adopted", "Not Adopted", "Adopted", "Not Adopted", "Adopted")
)

make_barplot(data, x = "animal_type", x_name = "Animal Type",
             class = "outcome_group", class_name = "Outcome Group")
}
}
