% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{make_table}
\alias{make_table}
\title{Creates a Table and Writes It to a CSV File}
\arguments{
\item{data_cols}{A dataframe or individual column(s) from a dataframe to create the table.}

\item{folder}{The folder and path where the CSV file will be saved.}

\item{name}{The name of the CSV file to be written.}
}
\value{
Writes a new CSV file to the specified folder.
}
\description{
Creates a table using the given data columns and writes it to the specified folder as a CSV file.
}
\examples{
# Example usage:
make_table(data_cols, folder = "path/to/folder", name = "output.csv")
}
